LayoutTests/platform/chromium/fast/url/segments-from-data-url-expected.txt

Test URL segmentation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS segments('http://user:pass@foo:21/bar;par?b#c') is '["http:","foo","21","/bar;par","?b","#c"]'
PASS segments('http:foo.com') is '["http:","foo.com","","/","",""]'
PASS segments('\t   :foo.com   \n') is '[":","","","","",""]'
PASS segments(' foo.com  ') is '[":","","","","",""]'
PASS segments('a:\t foo.com') is '["a:","",""," foo.com","",""]'
PASS segments('http://f:21/ b ? d # e ') is '["http:","f","21","/%20b%20","?%20d%20","# e"]'
PASS segments('http://f:/c') is '["http:","f","","/c","",""]'
PASS segments('http://f:0/c') is '["http:","f","0","/c","",""]'
PASS segments('http://f:00000000000000/c') is '["http:","f","0","/c","",""]'
FAIL segments('http://f:00000000000000000000080/c') should be ["http:","f","0","/c","",""]. Was ["http:","f","","/c","",""].
FAIL segments('http://f:b/c') should be [":","","","","",""]. Was [":","","0","","",""].
FAIL segments('http://f: /c') should be [":","","","","",""]. Was [":","","0","","",""].
FAIL segments('http://f:\n/c') should be [":","","","","",""]. Was ["http:","f","","/c","",""].
FAIL segments('http://f:fifty-two/c') should be [":","","","","",""]. Was [":","","0","","",""].
PASS segments('http://f:999999/c') is '[":","","0","","",""]'
FAIL segments('http://f: 21 / b ? d # e ') should be [":","","","","",""]. Was [":","","0","","",""].
PASS segments('') is '["data:","","","text/plain,baseURL","",""]'
PASS segments('  \t') is '["data:","","","text/plain,baseURL","",""]'
PASS segments(':foo.com/') is '[":","","","","",""]'
PASS segments(':foo.com\\') is '[":","","","","",""]'
PASS segments(':') is '[":","","","","",""]'
PASS segments(':a') is '[":","","","","",""]'
PASS segments(':/') is '[":","","","","",""]'
PASS segments(':\\') is '[":","","","","",""]'
PASS segments(':#') is '[":","","","","",""]'
PASS segments('#') is '[":","","","","",""]'
PASS segments('#/') is '[":","","","","",""]'
PASS segments('#\\') is '[":","","","","",""]'
PASS segments('#;?') is '[":","","","","",""]'
PASS segments('?') is '[":","","","","",""]'
PASS segments('/') is '[":","","","","",""]'
PASS segments(':23') is '[":","","","","",""]'
PASS segments('/:23') is '["data:","","","/:23","",""]'
PASS segments('//') is '[":","","","","",""]'
PASS segments('::') is '[":","","","","",""]'
PASS segments('::23') is '[":","","","","",""]'
PASS segments('foo://') is '["foo:","","","//","",""]'
PASS segments('http://a:b@c:29/d') is '["http:","c","29","/d","",""]'
PASS segments('http::@c:29') is '["http:","c","29","/","",""]'
PASS segments('http://&a:foo(b]c@d:2/') is '["http:","d","2","/","",""]'
PASS segments('http://::@c@d:2') is '["http:","d","2","/","",""]'
PASS segments('http://foo.com:b@d/') is '["http:","d","","/","",""]'
PASS segments('http://foo.com/\\@') is '["http:","foo.com","","//@","",""]'
PASS segments('http:\\\\foo.com\\') is '["http:","foo.com","","/","",""]'
PASS segments('http:\\\\a\\b:c\\d@foo.com\\') is '["http:","a","","/b:c/d@foo.com/","",""]'
PASS segments('foo:/') is '["foo:","","","/","",""]'
PASS segments('foo:/bar.com/') is '["foo:","","","/bar.com/","",""]'
PASS segments('foo://///////') is '["foo:","","","/////////","",""]'
PASS segments('foo://///////bar.com/') is '["foo:","","","/////////bar.com/","",""]'
PASS segments('foo:////://///') is '["foo:","","","////://///","",""]'
PASS segments('c:/foo') is '["c:","","","/foo","",""]'
PASS segments('//foo/bar') is '[":","","","","",""]'
PASS segments('http://foo/path;a??e#f#g') is '["http:","foo","","/path;a","??e","#f#g"]'
PASS segments('http://foo/abcd?efgh?ijkl') is '["http:","foo","","/abcd","?efgh?ijkl",""]'
PASS segments('http://foo/abcd#foo?bar') is '["http:","foo","","/abcd","","#foo?bar"]'
PASS segments('[61:24:74]:98') is '["data:","","","text/[61:24:74]:98","",""]'
PASS segments('http://[61:27]:98') is '[":","","0","","",""]'
PASS segments('http:[61:27]/:foo') is '[":","","","","",""]'
FAIL segments('http://[1::2]:3:4') should be [":","","","","",""]. Was [":","","0","","",""].
FAIL segments('http://2001::1') should be [":","","","","",""]. Was [":","","0","","",""].
PASS segments('http://[2001::1') is '[":","","","","",""]'
PASS segments('http://2001::1]') is '[":","","","","",""]'
PASS segments('http://2001::1]:80') is '[":","","","","",""]'
PASS segments('http://[2001::1]') is '["http:","[2001::1]","","/","",""]'
PASS segments('http://[2001::1]:80') is '["http:","[2001::1]","","/","",""]'
PASS segments('http://[[::]]') is '[":","","","","",""]'
PASS successfullyParsed is true

TEST COMPLETE
